Remove misplaced class
