--- flow-data/src/main/java/com/vaadin/data/provider/bov/Person.java
@@ -1,45 +1,0 @@
-/*
- * Copyright 2000-2017 Vaadin Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-package com.vaadin.data.provider.bov;
-
-import java.io.Serializable;
-
-public class Person implements Serializable {
-    private String name;
-    private final int born;
-
-    public Person(String name, int born) {
-        this.name = name;
-        this.born = born;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public int getBorn() {
-        return born;
-    }
-
-    @Override
-    public String toString() {
-        return name + "(" + born + ")";
-    }
-}
